+++ src/components/atoms/Box/index.stories.tsx
@@ -3,12 +3,12 @@ import { StoryFn, Meta } from '@storybook/react'
 // COMPONENTS
 import Box from '.'
 // MOCKS
-import mocks from './index.mocks.json'
+import { testing, storybook } from './index.mocks.json'
 
 export default {
   title: 'Atoms/Box',
   component: Box,
-  argTypes: mocks.storybook
+  argTypes: storybook
 } as Meta<typeof Box>
 
 const Template: StoryFn<typeof Box> = args => <Box {...args} />
@@ -19,7 +19,7 @@ BasicExample.storyName = 'No children'
 export const WithParagraph = Template.bind({})
 WithParagraph.storyName = 'With a paragraph'
 WithParagraph.args = {
-  children: <p>{mocks.testing.testParagraph}</p>
+  children: <p>{testing.testParagraph}</p>
 }
 
 export const OneBox = Template.bind({})
@@ -27,7 +27,7 @@ OneBox.storyName = 'Several paragraphs. One Box'
 OneBox.args = {
   children: Array(5)
     .fill(null)
-    .map((_, i) => <p key={`box-p-${i}`}>{mocks.testing.testParagraph}</p>)
+    .map((_, i) => <p key={`box-p-${i}`}>{testing.testParagraph}</p>)
 }
 
 export const SeveralBoxes = Template.bind({})
@@ -35,7 +35,5 @@ SeveralBoxes.storyName = 'Several paragraphs. Several Boxes'
 SeveralBoxes.args = {
   children: Array(5)
     .fill(null)
-    .map((_, i) => (
-      <Box key={`box-${i}`}>{<p>{mocks.testing.testParagraph}</p>}</Box>
-    ))
+    .map((_, i) => <Box key={`box-${i}`}>{<p>{testing.testParagraph}</p>}</Box>)
 }