+++ src/components/atoms/Checkbox/index.stories.tsx
@@ -1,27 +1,27 @@
 import React from 'react'
 import { StoryFn, Meta } from '@storybook/react'
 // COMPONENTS
-import Checkbox from '.'
+import CheckBox from '.'
+import Tag from '../Tag'
 // TYPES & INTERFACES
-import { CheckboxProps } from '../../../interfaces/atomProps'
+import { CheckBoxProps } from '../../../interfaces/atomProps'
 // MOCKS
 import { testing, storybook } from './index.mocks.json'
-import Tag from '../Tag'
 
 export default {
-  title: 'Atoms/Checkbox',
-  component: Checkbox,
+  title: 'Atoms/CheckBox',
+  component: CheckBox,
   argTypes: storybook
-} as Meta<typeof Checkbox>
+} as Meta<typeof CheckBox>
 
-const Template: StoryFn<typeof Checkbox> = args => <Checkbox {...args} />
+const Template: StoryFn<typeof CheckBox> = args => <CheckBox {...args} />
 
 export const BasicExample = Template.bind({})
 BasicExample.storyName = 'Basic Example'
 
 export const WithTextContent = Template.bind({})
 WithTextContent.storyName = 'With text content'
-WithTextContent.args = testing.withTextContent as CheckboxProps
+WithTextContent.args = testing.withTextContent as CheckBoxProps
 
 export const WithComponentContent = Template.bind({})
 WithComponentContent.storyName = 'With component content'
@@ -33,5 +33,5 @@ export const DisabledMode = Template.bind({})
 DisabledMode.storyName = 'isDisabled'
 DisabledMode.args = {
   ...WithTextContent.args,
-  ...(testing.disabledMode as CheckboxProps)
+  ...(testing.disabledMode as CheckBoxProps)
 }