+++ src/components/atoms/Checkbox/index.test.tsx
@@ -2,45 +2,46 @@ import React from 'react'
 import { fireEvent, render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
 // COMPONENTS
-import Checkbox from '.'
-// MOCKS
-import mocks from './index.mocks.json'
+import CheckBox from '.'
 import Tag from '../Tag'
+// MOCKS
+import { testing } from './index.mocks.json'
 
-describe('Checkbox', () => {
-  const { basicTestId, withTextContent, disabledMode } = mocks.testing
+describe('CheckBox', () => {
+  const { basicTestId, withTextContent, disabledMode } = testing
+
   test('Should render with minimal config', () => {
-    render(<Checkbox />)
-    const minimalTestCheckbox = screen.getByTestId(basicTestId)
-    expect(minimalTestCheckbox).toBeInTheDocument()
+    render(<CheckBox />)
+    const minimalTestCheckBox = screen.getByTestId(basicTestId)
+    expect(minimalTestCheckBox).toBeInTheDocument()
   })
 
   test('Should render with included text content', () => {
-    render(<Checkbox {...withTextContent} />)
-    const testTextContentCheckbox = screen.getByTestId(
+    render(<CheckBox {...withTextContent} />)
+    const testTextContentCheckBox = screen.getByTestId(
       `${basicTestId}-${withTextContent.content}`
     )
-    expect(testTextContentCheckbox).toBeInTheDocument()
+    expect(testTextContentCheckBox).toBeInTheDocument()
   })
 
   test('Should render with included text content', () => {
     const componentContentConfig = {
       content: <Tag text='test' />
     }
-    render(<Checkbox {...componentContentConfig} />)
-    const testTextContentCheckbox = screen.getByTestId(
+    render(<CheckBox {...componentContentConfig} />)
+    const testTextContentCheckBox = screen.getByTestId(
       `${basicTestId}-with-component`
     )
-    expect(testTextContentCheckbox).toBeInTheDocument()
+    expect(testTextContentCheckBox).toBeInTheDocument()
   })
 
   test('Should check that the checkbox has been clicked', () => {
     const clickeableConfig = { onChange: jest.fn() }
 
-    render(<Checkbox {...clickeableConfig} />)
-    const clickCheckbox = screen.getByTestId(basicTestId)
+    render(<CheckBox {...clickeableConfig} />)
+    const clickCheckBox = screen.getByTestId(basicTestId)
 
-    fireEvent.click(clickCheckbox)
+    fireEvent.click(clickCheckBox)
     expect(clickeableConfig.onChange).toHaveBeenCalled()
     expect(clickeableConfig.onChange).toHaveBeenCalledTimes(1)
   })
@@ -51,10 +52,10 @@ describe('Checkbox', () => {
       onChange: jest.fn()
     }
 
-    render(<Checkbox {...notClickeableConfig} />)
-    const disabledCheckbox = screen.getByTestId(basicTestId)
+    render(<CheckBox {...notClickeableConfig} />)
+    const disabledCheckBox = screen.getByTestId(basicTestId)
 
-    fireEvent.change(disabledCheckbox)
+    fireEvent.change(disabledCheckBox)
     expect(notClickeableConfig.onChange).not.toHaveBeenCalled()
     expect(notClickeableConfig.onChange).toHaveBeenCalledTimes(0)
   })