+++ src/components/atoms/Delete/index.stories.tsx
@@ -5,12 +5,11 @@ import Delete from '.'
 // TYPES & INTERFACES
 import { IconProps } from '../../../interfaces/atomProps'
 // MOCKS
-import mocks from './index.mocks.json'
+import { testing } from './index.mocks.json'
 
 export default {
   title: 'Atoms/Delete',
-  component: Delete,
-  argTypes: mocks.storybook
+  component: Delete
 } as Meta<typeof Delete>
 
 const Template: StoryFn<typeof Delete> = args => <Delete {...args} />
@@ -20,4 +19,4 @@ BasicExample.storyName = 'Basic Example'
 
 export const BigSize = Template.bind({})
 BigSize.storyName = 'With Big Size'
-BigSize.args = mocks.testing.bigSize as IconProps
+BigSize.args = testing.bigSize as IconProps