+++ src/components/atoms/Select/index.stories.tsx
@@ -3,12 +3,12 @@ import { StoryFn, Meta } from '@storybook/react'
 // COMPONENTS
 import Select from '.'
 // MOCKS
-import mocks from './index.mocks.json'
+import { testing, storybook } from './index.mocks.json'
 
 export default {
   title: 'Atoms/Select',
   component: Select,
-  argTypes: mocks.storybook
+  argTypes: storybook
 } as Meta<typeof Select>
 
 const Template: StoryFn<typeof Select> = args => <Select {...args} />
@@ -18,22 +18,22 @@ OnlyRequiredProps.storyName = 'Only Required props'
 
 export const WithOneOption = Template.bind({})
 WithOneOption.storyName = 'With 1 option'
-WithOneOption.args = mocks.testing.oneOption
+WithOneOption.args = testing.oneOption
 
 export const WithSeveralOptions = Template.bind({})
 WithSeveralOptions.storyName = 'With several options'
-WithSeveralOptions.args = mocks.testing.severalOptions
+WithSeveralOptions.args = testing.severalOptions
 
 export const ShowsSeveralOptions = Template.bind({})
 ShowsSeveralOptions.storyName = 'With several options'
 ShowsSeveralOptions.args = {
-  ...mocks.testing.severalOptions,
-  ...mocks.testing.showsThreeOptions
+  ...testing.severalOptions,
+  ...testing.showsThreeOptions
 }
 
 export const FirstOptionsSelected = Template.bind({})
 FirstOptionsSelected.storyName = 'First option selected'
 FirstOptionsSelected.args = {
-  ...mocks.testing.showsThreeOptions,
-  ...mocks.testing.firstOptionSelected
+  ...testing.showsThreeOptions,
+  ...testing.firstOptionSelected
 }